cleanup the unsigned types on the api to make it simpler

+++ src/androidMain/kotlin/com/ditchoom/buffer/BaseJvmBuffer.kt
@@ -30,10 +30,10 @@ abstract class BaseJvmBuffer(val byteBuffer: ByteBuffer, val fileRef: RandomAcce
         buffer.limit(limit)
     }
 
-    override val capacity = buffer.capacity().toUInt()
+    override val capacity = buffer.capacity()
 
     override fun readByte() = byteBuffer.get()
-    override fun readByteArray(size: UInt) = byteBuffer.toArray(size)
+    override fun readByteArray(size: Int) = byteBuffer.toArray(size)
 
     override fun slice() = JvmBuffer(byteBuffer.slice())
 
@@ -44,8 +44,8 @@ abstract class BaseJvmBuffer(val byteBuffer: ByteBuffer, val fileRef: RandomAcce
     override fun readUnsignedInt() = byteBuffer.int.toUInt()
     override fun readLong() = byteBuffer.long
 
-    override fun readUtf8(bytes: UInt): CharSequence {
-        val finalPosition = buffer.position() + bytes.toInt()
+    override fun readUtf8(bytes: Int): CharSequence {
+        val finalPosition = buffer.position() + bytes
         val readBuffer = byteBuffer.asReadOnlyBuffer()
         (readBuffer as Buffer).limit(finalPosition)
         val decoded = Charsets.UTF_8.decode(readBuffer)
@@ -114,8 +114,8 @@ abstract class BaseJvmBuffer(val byteBuffer: ByteBuffer, val fileRef: RandomAcce
         fileRef?.aClose()
     }
 
-    override fun limit() = buffer.limit().toUInt()
-    override fun position() = buffer.position().toUInt()
+    override fun limit() = buffer.limit()
+    override fun position() = buffer.position()
 }
 
 
@@ -131,13 +131,13 @@ suspend fun RandomAccessFile.aClose() = suspendCoroutine<Unit> {
 }
 
 
-fun ByteBuffer.toArray(size: UInt = remaining().toUInt()): ByteArray {
+fun ByteBuffer.toArray(size: Int = remaining()): ByteArray {
     return if (hasArray()) {
-        val result = ByteArray(size.toInt())
-        System.arraycopy(this.array(), position(), result, 0, size.toInt())
+        val result = ByteArray(size)
+        System.arraycopy(this.array(), position(), result, 0, size)
         result
     } else {
-        val byteArray = ByteArray(size.toInt())
+        val byteArray = ByteArray(size)
         get(byteArray)
         byteArray
     }
