+++ src/commonTest/kotlin/com/ditchoom/buffer/BufferTests.kt
@@ -12,15 +12,15 @@ class BufferTests {
 
     @Test
     fun slice() {
-        val platformBuffer = PlatformBuffer.allocate(3u)
+        val platformBuffer = PlatformBuffer.allocate(3)
         platformBuffer.write((-1).toByte())
         platformBuffer.resetForRead()
         val slicedBuffer = platformBuffer.slice()
-        assertEquals(0u, slicedBuffer.position())
-        assertEquals(1u, slicedBuffer.limit())
+        assertEquals(0, slicedBuffer.position())
+        assertEquals(1, slicedBuffer.limit())
         assertEquals(-1, slicedBuffer.readByte())
-        assertEquals(1u, slicedBuffer.position())
-        assertEquals(1u, slicedBuffer.limit())
+        assertEquals(1, slicedBuffer.position())
+        assertEquals(1, slicedBuffer.limit())
     }
 
 
@@ -33,7 +33,7 @@ class BufferTests {
             expected[1].code.toByte(),
             expected[2].code.toByte(),
             expected[3].code.toByte())
-        val platformBuffer = PlatformBuffer.allocate(bytes.size.toUInt())
+        val platformBuffer = PlatformBuffer.allocate(bytes.size)
         platformBuffer.write(bytes)
         platformBuffer.position(2)
         assertEquals(expected, platformBuffer.readUtf8(4).toString())
@@ -41,27 +41,27 @@ class BufferTests {
 
     @Test
     fun sliceFragmented() {
-        val platformBuffer1 = PlatformBuffer.allocate(3u)
+        val platformBuffer1 = PlatformBuffer.allocate(3)
         platformBuffer1.write(1.toByte())
         platformBuffer1.resetForRead()
 
-        val platformBuffer2 = PlatformBuffer.allocate(3u)
+        val platformBuffer2 = PlatformBuffer.allocate(3)
         platformBuffer2.write(2.toByte())
         platformBuffer2.resetForRead()
 
         val fragmentedBuffer = FragmentedReadBuffer(platformBuffer1, platformBuffer2)
         val slicedBuffer = fragmentedBuffer.slice()
-        assertEquals(0u, slicedBuffer.position())
-        assertEquals(2u, slicedBuffer.limit())
+        assertEquals(0, slicedBuffer.position())
+        assertEquals(2, slicedBuffer.limit())
         assertEquals(1, slicedBuffer.readByte())
         assertEquals(2, slicedBuffer.readByte())
-        assertEquals(2u, slicedBuffer.position())
-        assertEquals(2u, slicedBuffer.limit())
+        assertEquals(2, slicedBuffer.position())
+        assertEquals(2, slicedBuffer.limit())
     }
 
     @Test
     fun byte() {
-        val platformBuffer = PlatformBuffer.allocate(1u)
+        val platformBuffer = PlatformBuffer.allocate(1)
         val byte = (-1).toByte()
         platformBuffer.write(byte)
         platformBuffer.resetForRead()
@@ -71,11 +71,11 @@ class BufferTests {
     @Test
     fun byteArray() {
         val size = 200
-        val platformBuffer = PlatformBuffer.allocate(size.toUInt())
+        val platformBuffer = PlatformBuffer.allocate(size)
         val bytes = ByteArray(200) { -1 }
         platformBuffer.write(bytes)
         platformBuffer.resetForRead()
-        val byteArray = platformBuffer.readByteArray(size.toUInt())
+        val byteArray = platformBuffer.readByteArray(size)
         assertEquals(bytes.count(), byteArray.count())
         var count = 0
         for (byte in bytes) {
@@ -85,7 +85,7 @@ class BufferTests {
 
     @Test
     fun unsignedByte() {
-        val platformBuffer = PlatformBuffer.allocate(1u)
+        val platformBuffer = PlatformBuffer.allocate(1)
         val byte = (-1).toUByte()
         platformBuffer.write(byte)
         platformBuffer.resetForRead()
@@ -94,7 +94,7 @@ class BufferTests {
 
     @Test
     fun unsignedShort() {
-        val platformBuffer = PlatformBuffer.allocate(2u)
+        val platformBuffer = PlatformBuffer.allocate(2)
         val uShort = UShort.MAX_VALUE.toInt() / 2
         platformBuffer.write(uShort.toUShort())
         platformBuffer.resetForRead()
@@ -109,7 +109,7 @@ class BufferTests {
 
     @Test
     fun allUShortValues() {
-        val buffer = PlatformBuffer.allocate(UShort.MAX_VALUE.toUInt() * UShort.SIZE_BYTES.toUInt())
+        val buffer = PlatformBuffer.allocate(UShort.MAX_VALUE.toInt() * UShort.SIZE_BYTES)
         (0 until UShort.MAX_VALUE.toInt()).forEach {
             buffer.write(it.toUShort())
         }
@@ -121,7 +121,7 @@ class BufferTests {
 
     @Test
     fun unsignedShortHalf() {
-        val platformBuffer = PlatformBuffer.allocate(2u)
+        val platformBuffer = PlatformBuffer.allocate(2)
         val uShort = (UShort.MAX_VALUE / 2u).toUShort()
         platformBuffer.write(uShort)
         platformBuffer.resetForRead()
@@ -132,7 +132,7 @@ class BufferTests {
 
     @Test
     fun unsignedInt() {
-        val platformBuffer = PlatformBuffer.allocate(4u)
+        val platformBuffer = PlatformBuffer.allocate(4)
         val uInt = (-1).toUInt()
         platformBuffer.write(uInt)
         platformBuffer.resetForRead()
@@ -141,7 +141,7 @@ class BufferTests {
 
     @Test
     fun unsignedIntHalf() {
-        val platformBuffer = PlatformBuffer.allocate(4u)
+        val platformBuffer = PlatformBuffer.allocate(4)
         val uInt = Int.MAX_VALUE.toUInt() / 2u
         platformBuffer.write(uInt)
         platformBuffer.resetForRead()
@@ -150,7 +150,7 @@ class BufferTests {
 
     @Test
     fun long() {
-        val platformBuffer = PlatformBuffer.allocate(Long.SIZE_BYTES.toUInt())
+        val platformBuffer = PlatformBuffer.allocate(Long.SIZE_BYTES)
         val long = (1234).toLong()
         platformBuffer.write(long)
         platformBuffer.resetForRead()
@@ -159,7 +159,7 @@ class BufferTests {
 
     @Test
     fun float() {
-        val platformBuffer = PlatformBuffer.allocate(Float.SIZE_BYTES.toUInt())
+        val platformBuffer = PlatformBuffer.allocate(Float.SIZE_BYTES)
         val float = 123.456f
         platformBuffer.write(float)
         platformBuffer.resetForRead()
@@ -170,7 +170,7 @@ class BufferTests {
 
     @Test
     fun double() {
-        val platformBuffer = PlatformBuffer.allocate(Double.SIZE_BYTES.toUInt())
+        val platformBuffer = PlatformBuffer.allocate(Double.SIZE_BYTES)
         val double = 123.456
         platformBuffer.write(double)
         platformBuffer.resetForRead()
@@ -180,8 +180,8 @@ class BufferTests {
     @Test
     fun utf8String() {
         val string = "yolo swag lyfestyle"
-        assertEquals(19, string.utf8Length().toInt())
-        val platformBuffer = PlatformBuffer.allocate(19u)
+        assertEquals(19, string.utf8Length())
+        val platformBuffer = PlatformBuffer.allocate(19)
         platformBuffer.writeUtf8(string)
         platformBuffer.resetForRead()
         val actual = platformBuffer.readUtf8(19u).toString()
@@ -195,7 +195,7 @@ class BufferTests {
         val text = "hello"
         val buffer = text.toBuffer()
         assertEquals("hello", buffer.readUtf8Line().toString())
-        assertEquals(buffer.remaining(), 0u)
+        assertEquals(buffer.remaining(), 0)
     }
 
     @Test
@@ -204,7 +204,7 @@ class BufferTests {
         val buffer = text.toBuffer()
         assertEquals("hello", buffer.readUtf8Line().toString())
         assertEquals("", buffer.readUtf8Line().toString())
-        assertEquals(buffer.remaining(), 0u)
+        assertEquals(buffer.remaining(), 0)
     }
 
     @Test
@@ -213,7 +213,7 @@ class BufferTests {
         val buffer = text.toBuffer()
         assertEquals("", buffer.readUtf8Line().toString())
         assertEquals("hello", buffer.readUtf8Line().toString())
-        assertEquals(buffer.remaining(), 0u)
+        assertEquals(buffer.remaining(), 0)
     }
 
     @Test
@@ -222,7 +222,7 @@ class BufferTests {
         val buffer = text.toBuffer()
         assertEquals("", buffer.readUtf8Line().toString())
         assertEquals("hello", buffer.readUtf8Line().toString())
-        assertEquals(buffer.remaining(), 0u)
+        assertEquals(buffer.remaining(), 0)
     }
 
     @Test
@@ -234,7 +234,7 @@ class BufferTests {
         assertEquals("hello", buffer.readUtf8Line().toString())
         assertEquals("hello", buffer.readUtf8Line().toString())
         assertEquals("", buffer.readUtf8Line().toString())
-        assertEquals(buffer.remaining(), 0u)
+        assertEquals(buffer.remaining(), 0)
     }
 
 
@@ -249,7 +249,7 @@ class BufferTests {
         assertEquals("", buffer.readUtf8Line().toString())
         assertEquals("hello", buffer.readUtf8Line().toString())
         assertEquals("", buffer.readUtf8Line().toString())
-        assertEquals(buffer.remaining(), 0u)
+        assertEquals(buffer.remaining(), 0)
     }
 
     @Test
@@ -266,19 +266,19 @@ class BufferTests {
 
     @Test
     fun endianWrite() {
-        val littleEndian2 = PlatformBuffer.allocate(2u, ByteOrder.LITTLE_ENDIAN)
+        val littleEndian2 = PlatformBuffer.allocate(2, ByteOrder.LITTLE_ENDIAN)
         littleEndian2.write(0x0102.toShort())
         littleEndian2.resetForRead()
         assertEquals(0x02u, littleEndian2.readUnsignedByte())
         assertEquals(0x01u, littleEndian2.readUnsignedByte())
 
-        val bigEndian2 = PlatformBuffer.allocate(2u, ByteOrder.BIG_ENDIAN)
+        val bigEndian2 = PlatformBuffer.allocate(2, ByteOrder.BIG_ENDIAN)
         bigEndian2.write(0x0102.toShort())
         bigEndian2.resetForRead()
         assertEquals(0x01u, bigEndian2.readUnsignedByte())
         assertEquals(0x02u, bigEndian2.readUnsignedByte())
 
-        val littleEndian4 = PlatformBuffer.allocate(4u, ByteOrder.LITTLE_ENDIAN)
+        val littleEndian4 = PlatformBuffer.allocate(4, ByteOrder.LITTLE_ENDIAN)
         littleEndian4.write(0x01020304)
         littleEndian4.resetForRead()
         assertEquals(0x04u, littleEndian4.readUnsignedByte())
@@ -286,7 +286,7 @@ class BufferTests {
         assertEquals(0x02u, littleEndian4.readUnsignedByte())
         assertEquals(0x01u, littleEndian4.readUnsignedByte())
 
-        val bigEndian4 = PlatformBuffer.allocate(4u, ByteOrder.BIG_ENDIAN)
+        val bigEndian4 = PlatformBuffer.allocate(4, ByteOrder.BIG_ENDIAN)
         bigEndian4.write(0x01020304)
         bigEndian4.resetForRead()
         assertEquals(0x01u, bigEndian4.readUnsignedByte())
@@ -294,7 +294,7 @@ class BufferTests {
         assertEquals(0x03u, bigEndian4.readUnsignedByte())
         assertEquals(0x04u, bigEndian4.readUnsignedByte())
 
-        val littleEndian8 = PlatformBuffer.allocate(8u, ByteOrder.LITTLE_ENDIAN)
+        val littleEndian8 = PlatformBuffer.allocate(8, ByteOrder.LITTLE_ENDIAN)
         littleEndian8.write(0x0102030405060708)
         littleEndian8.resetForRead()
         assertEquals(0x08u, littleEndian8.readUnsignedByte())
@@ -306,7 +306,7 @@ class BufferTests {
         assertEquals(0x02u, littleEndian8.readUnsignedByte())
         assertEquals(0x01u, littleEndian8.readUnsignedByte())
 
-        val bigEndian8 = PlatformBuffer.allocate(8u, ByteOrder.BIG_ENDIAN)
+        val bigEndian8 = PlatformBuffer.allocate(8, ByteOrder.BIG_ENDIAN)
         bigEndian8.write(0x0102030405060708)
         bigEndian8.resetForRead()
         assertEquals(0x01u, bigEndian8.readUnsignedByte())
@@ -321,19 +321,19 @@ class BufferTests {
 
     @Test
     fun endianRead() {
-        val littleEndian2 = PlatformBuffer.allocate(2u, ByteOrder.LITTLE_ENDIAN)
+        val littleEndian2 = PlatformBuffer.allocate(2, ByteOrder.LITTLE_ENDIAN)
         littleEndian2.write(0x01.toByte())
         littleEndian2.write(0x02.toByte())
         littleEndian2.resetForRead()
         assertEquals(0x0201.toShort(), littleEndian2.readShort())
 
-        val bigEndian2 = PlatformBuffer.allocate(2u, ByteOrder.BIG_ENDIAN)
+        val bigEndian2 = PlatformBuffer.allocate(2, ByteOrder.BIG_ENDIAN)
         bigEndian2.write(0x01.toByte())
         bigEndian2.write(0x02.toByte())
         bigEndian2.resetForRead()
         assertEquals(0x0102.toShort(), bigEndian2.readShort())
 
-        val littleEndian4 = PlatformBuffer.allocate(4u, ByteOrder.LITTLE_ENDIAN)
+        val littleEndian4 = PlatformBuffer.allocate(4, ByteOrder.LITTLE_ENDIAN)
         littleEndian4.write(0x01.toByte())
         littleEndian4.write(0x02.toByte())
         littleEndian4.write(0x03.toByte())
@@ -341,7 +341,7 @@ class BufferTests {
         littleEndian4.resetForRead()
         assertEquals(0x04030201, littleEndian4.readInt())
 
-        val bigEndian4 = PlatformBuffer.allocate(4u, ByteOrder.BIG_ENDIAN)
+        val bigEndian4 = PlatformBuffer.allocate(4, ByteOrder.BIG_ENDIAN)
         bigEndian4.write(0x01.toByte())
         bigEndian4.write(0x02.toByte())
         bigEndian4.write(0x03.toByte())
@@ -349,7 +349,7 @@ class BufferTests {
         bigEndian4.resetForRead()
         assertEquals(0x01020304, bigEndian4.readInt())
 
-        val littleEndian8 = PlatformBuffer.allocate(8u, ByteOrder.LITTLE_ENDIAN)
+        val littleEndian8 = PlatformBuffer.allocate(8, ByteOrder.LITTLE_ENDIAN)
         littleEndian8.write(0x01.toByte())
         littleEndian8.write(0x02.toByte())
         littleEndian8.write(0x03.toByte())
@@ -361,7 +361,7 @@ class BufferTests {
         littleEndian8.resetForRead()
         assertEquals(0x0807060504030201, littleEndian8.readLong())
 
-        val bigEndian8 = PlatformBuffer.allocate(8u, ByteOrder.BIG_ENDIAN)
+        val bigEndian8 = PlatformBuffer.allocate(8, ByteOrder.BIG_ENDIAN)
         bigEndian8.write(0x01.toByte())
         bigEndian8.write(0x02.toByte())
         bigEndian8.write(0x03.toByte())
@@ -378,17 +378,17 @@ class BufferTests {
     fun partialByteArray() {
         val byteArray = byteArrayOf(0,1,2,3,4,5,6,7,8,9)
         val partialArray = byteArray.sliceArray(2..6)
-        val buffer = PlatformBuffer.allocate(partialArray.size.toUInt())
+        val buffer = PlatformBuffer.allocate(partialArray.size)
         buffer.write(byteArray, 2, 5)
         buffer.resetForRead()
-        assertContentEquals(partialArray, buffer.readByteArray(5u))
+        assertContentEquals(partialArray, buffer.readByteArray(5))
     }
 
     @Test
     fun wrap() {
         val byteArray = byteArrayOf(0,1,2,3,4,5,6,7,8,9)
         val buffer = PlatformBuffer.wrap(byteArray)
-        assertEquals(byteArray.size.toUInt(), buffer.remaining())
+        assertEquals(byteArray.size, buffer.remaining())
         assertContentEquals(byteArray, buffer.readByteArray(buffer.remaining()))
     }
 }
